Add Genserver to fetch and cache credentials
This commit adds a GenServer that starts `AwsCredentials` application
with a specified provider. This will fetch AWS credentials and cache
them.

If the credentials have expiration (in case of EC2 provider) then
whenever the next time credentials are fetched the cache is updated.

Expiration is determined by the simple formula: if current time and
expiration time difference is >= expiration threshold. Default threshold
is 10 minutes.